chore#11: moved debugging logic to separate file to simplify ada.go
